FIX(SearchScreen): 뒤로 가기 버튼 동작 수정
