+++ presentation/src/main/java/com/capston/presentation/ui/MainActivity.kt
@@ -279,76 +279,6 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun SearchFieldWithIcons(
-    searchQuery: String,
-    onQueryChanged: (String) -> Unit,
-    onBackClick: () -> Unit,
-    onSearchClick: () -> Unit
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(end = 16.dp)
-            .height(40.dp)
-            .border(BorderStroke(0.5.dp, Color.LightGray), shape = RoundedCornerShape(20.dp))
-            .background(LightGray4_40, shape = RoundedCornerShape(20.dp))
-            .padding(horizontal = 8.dp)
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.fillMaxSize() // Use fillMaxSize to take up the full height
-        ) {
-            IconButton(onClick = onBackClick, modifier = Modifier.size(24.dp)) {
-                Icon(imageVector = Icons.Default.ArrowBack, contentDescription = "뒤로 가기", tint = Color.Gray)
-            }
-
-            BasicTextField(
-                value = searchQuery,
-                onValueChange = onQueryChanged,
-                singleLine = true,
-                textStyle = TextStyle(fontSize = 14.sp, color = Color.DarkGray),
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(horizontal = 8.dp)
-                    .wrapContentHeight(Alignment.CenterVertically), // Center text vertically
-                decorationBox = { innerTextField ->
-                    Box(
-                        contentAlignment = Alignment.CenterStart,
-                        modifier = Modifier.fillMaxHeight()
-                    ) {
-                        if (searchQuery.isEmpty()) {
-                            Text(
-                                text = "계획을 생성하고 싶은 강의 또는 선생님을 검색하세요",
-                                fontSize = 12.sp,
-                                color = Color.Gray
-                            )
-                        }
-                        innerTextField()
-                    }
-                }
-            )
-
-            if (searchQuery.isNotEmpty()) {
-                IconButton(
-                    onClick = { onQueryChanged("") },
-                    modifier = Modifier.size(20.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Close,
-                        contentDescription = "Clear",
-                        tint = Color.Gray
-                    )
-                }
-            }
-
-            IconButton(onClick = { onSearchClick() }, modifier = Modifier.size(24.dp)) {
-                Icon(imageVector = Icons.Default.Search, contentDescription = "검색", tint = Color.Gray)
-            }
-        }
-    }
-}
-
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun MainBottomBar(
