+++ presentation/src/main/java/com/capston/presentation/ui/search/SearchScreen.kt
@@ -1,6 +1,7 @@
 package com.capston.presentation.ui.search
 
 import android.annotation.SuppressLint
+import android.app.Activity
 import android.content.Intent
 import android.os.Build
 import android.util.Log
@@ -19,6 +20,7 @@ import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -28,11 +30,13 @@ import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Close
@@ -42,10 +46,9 @@ import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
-import androidx.compose.material3.SuggestionChip
-import androidx.compose.material3.SuggestionChipDefaults
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
@@ -69,6 +72,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
@@ -96,10 +100,9 @@ import com.capston.presentation.theme.MainPurple
 import com.capston.presentation.theme.materialGray
 import com.capston.presentation.theme.textGray
 import com.capston.domain.model.LectureItemDto
-import com.capston.presentation.ui.common.LoadingIndicator
+import com.capston.presentation.theme.LightGray4_40
 import com.capston.presentation.ui.MainActivity
 import com.capston.presentation.ui.common.Screen
-import com.capston.presentation.ui.SearchFieldWithIcons
 import com.capston.presentation.viewmodel.LectureViewModel
 import com.capston.presentation.viewmodel.PlanViewModel
 import kotlinx.coroutines.Job
@@ -115,6 +118,7 @@ fun SearchScreen(
     initialQuery: String = "",
     loadingStateManager: LoadingStateManager
 ) {
+    val context = LocalContext.current
     var searchQuery by remember { mutableStateOf(initialQuery) }
     val scope = rememberCoroutineScope()
     val searchLectureResponse by lectureViewModel.distinctLecture.collectAsState()
@@ -410,7 +414,11 @@ fun SearchScreen(
                 searchQuery = searchQuery,
                 onQueryChanged = { searchQuery = it },
                 lectureViewModel = lectureViewModel,
-                onSearchClick = {}
+                onSearchClick = {},
+                onBackClick = {
+                    // 현재 컨텍스트가 Activity인지 확인하고 finish() 호출
+                    (context as? Activity)?.finish()
+                },
             )
 
             // 필터 바 컴포넌트
@@ -1013,6 +1021,7 @@ fun SearchTopBar(
     searchQuery: String,
     onQueryChanged: (String) -> Unit,
     lectureViewModel: LectureViewModel,
+    onBackClick: () -> Unit,
     onSearchClick: () -> Unit
 ) {
     val context = LocalContext.current
@@ -1022,19 +1031,100 @@ fun SearchTopBar(
             SearchFieldWithIcons(
                 searchQuery = searchQuery,
                 onQueryChanged = onQueryChanged,
-                onBackClick = {
-                    val intent = Intent(context, MainActivity::class.java)
-                    context.startActivity(intent)
-                },
-                onSearchClick = {
-                    // 검색 버튼을 클릭했을 때 검색 실행
-                    onSearchClick()
-                }
+                onBackClick = onBackClick,
+                onSearchClick = onSearchClick
             )
         }
     )
 }
 
+@Composable
+fun SearchFieldWithIcons(
+    searchQuery: String,
+    onQueryChanged: (String) -> Unit,
+    onBackClick: () -> Unit,
+    onSearchClick: () -> Unit
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(end = 16.dp)
+            .height(40.dp)
+            .border(BorderStroke(0.5.dp, Color.LightGray), shape = RoundedCornerShape(20.dp))
+            .background(LightGray4_40, shape = RoundedCornerShape(20.dp))
+            .padding(horizontal = 8.dp)
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier.fillMaxSize() // Use fillMaxSize to take up the full height
+        ) {
+            IconButton(
+                onClick = onBackClick,
+                modifier = Modifier
+                    .size(24.dp)
+                    .padding(start = 4.dp)
+            ) {
+                Icon(
+                    painter = painterResource(R.drawable.icon_arrow_back),
+                    contentDescription = "뒤로 가기",
+                    tint = Color.Gray
+                )
+            }
+
+            BasicTextField(
+                value = searchQuery,
+                onValueChange = onQueryChanged,
+                singleLine = true,
+                textStyle = TextStyle(fontSize = 14.sp, color = Color.DarkGray),
+                modifier = Modifier
+                    .weight(1f)
+                    .padding(horizontal = 8.dp)
+                    .wrapContentHeight(Alignment.CenterVertically), // Center text vertically
+                decorationBox = { innerTextField ->
+                    Box(
+                        contentAlignment = Alignment.CenterStart,
+                        modifier = Modifier.fillMaxHeight()
+                    ) {
+                        if (searchQuery.isEmpty()) {
+                            Text(
+                                text = "계획을 생성하고 싶은 강의 또는 선생님을 검색하세요",
+                                style = MaterialTheme.typography.labelSmall,
+                                color = Color.Gray
+                            )
+                        }
+                        innerTextField()
+                    }
+                }
+            )
+
+            if (searchQuery.isNotEmpty()) {
+                IconButton(
+                    onClick = { onQueryChanged("") },
+                    modifier = Modifier.size(20.dp)
+                ) {
+                    Icon(
+                        painter = painterResource(R.drawable.icon_xmark),
+                        contentDescription = "Clear",
+                        tint = Color.Gray
+                    )
+                }
+            }
+
+            IconButton(
+                onClick = { onSearchClick() },
+                modifier = Modifier
+                    .size(24.dp)
+                    .padding(end = 4.dp)
+            ) {
+                Icon(
+                    painter = painterResource(R.drawable.icon_search),
+                    contentDescription = "검색",
+                    tint = Color.Gray
+                )
+            }
+        }
+    }
+}
 
 val Subject.bgColor: Color
     get() = when (this) {
